+++ options.c
@@ -1413,12 +1413,19 @@ int parse_arguments(int *argc_p, const char ***argv_p)
 			break;
 
 		case OPT_DAEMON:
+#ifdef WIN32_NATIVE
+			rprintf(FERROR,
+				"--daemon, --config, --dparam, and --no-detach are not supported "
+				"in this Windows build.\nUse rsync-over-ssh instead.\n");
+			exit_cleanup(RERR_UNSUPPORTED);
+#else
 			if (am_daemon) {
 				strlcpy(err_buf,
 					"Attempt to hack rsync thwarted!\n",
 					sizeof err_buf);
 				goto cleanup;
 			}
+#endif
 #ifdef ICONV_OPTION
 			iconv_opt = NULL;
 #endif
@@ -3140,6 +3147,12 @@ char *check_for_hostspec(char *s, char **host_ptr, int *port_ptr)
 #endif
 
 	if (port_ptr && strncasecmp(URL_PREFIX, s, strlen(URL_PREFIX)) == 0) {
+#ifdef WIN32_NATIVE
+		rprintf(FERROR,
+			"rsync:// URLs (daemon connections) are not supported in this Windows build.\n"
+			"Use rsync-over-ssh: rsync user@host:/path/ destination/\n");
+		exit_cleanup(RERR_UNSUPPORTED);
+#endif
 		*host_ptr = parse_hostspec(s + strlen(URL_PREFIX), &path, port_ptr);
 		if (*host_ptr) {
 			if (!*port_ptr)